components/voiceAgent/livekit-agent/requirements.txt
livekit-agents>=1.3.0
livekit-plugins-openai>=1.3.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
